+++ Packages/Tracking/Examples~/Shared Example Assets REQUIRED/Scripts/CycleHandPairs.cs
@@ -1,13 +1,16 @@
 /******************************************************************************
- * Copyright (C) Ultraleap, Inc. 2011-2024.                                   *
+ * Copyright (C) Ultraleap, Inc. 2011-2025.                                   *
  *                                                                            *
  * Use subject to the terms of the Apache License 2.0 available at            *
  * http://www.apache.org/licenses/LICENSE-2.0, or another agreement           *
  * between Ultraleap and you, your company or other organization.             *
  ******************************************************************************/
 
-using System.Collections.Generic;
 using UnityEngine;
+
+#if UNITY_INPUT_SYSTEM
+using UnityEngine.InputSystem;
+#endif
 
 namespace Leap.HandsModule.Examples
 {
@@ -41,6 +44,19 @@ namespace Leap.HandsModule.Examples
                 return;
             }
 
+#if UNITY_INPUT_SYSTEM
+            if (Keyboard.current != null)
+            {
+                if (Keyboard.current.rightArrowKey.wasReleasedThisFrame)
+                {
+                    NextHandSet();
+                }
+                if (Keyboard.current.leftArrowKey.wasReleasedThisFrame)
+                {
+                    PreviousHandSet();
+                }
+            }
+#else
             if (Input.GetKeyUp(KeyCode.RightArrow))
             {
                 NextHandSet();
@@ -50,6 +66,7 @@ namespace Leap.HandsModule.Examples
             {
                 PreviousHandSet();
             }
+#endif
         }
 
         private void NextHandSet()